remove old gallery
