+++ sveltekit-app/src/routes/gallery/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { PageData } from './$types';
-	import Gallery from '$lib/components/Gallery2.svelte';
+	import Gallery from '$lib/components/Gallery.svelte';
 
 	export let data: PageData;
